feat: give every Vue host the one write path, and the store one writer
