+++ common/changes/@stonecrop/desktop/action-payload-re-export_2026-08-11-14-00-00.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "comment": "Re-export ActionEventPayload from @stonecrop/stonecrop, so the shell that emits it and the composable that consumes it can no longer drift apart.",
+      "type": "patch",
+      "packageName": "@stonecrop/desktop"
+    }
+  ],
+  "packageName": "@stonecrop/desktop"
+}
