+++ common/changes/@stonecrop/nuxt/client-action-re-export_2026-08-11-14-00-00.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "comment": "Re-export useClientAction from @stonecrop/stonecrop so every Vue host can reach it; the Nuxt auto-import and existing @action bindings are unchanged.",
+      "type": "patch",
+      "packageName": "@stonecrop/nuxt"
+    }
+  ],
+  "packageName": "@stonecrop/nuxt"
+}
