+++ common/changes/@stonecrop/stonecrop/one-write-path-for-every-vue-host_2026-08-11-14-00-00.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "comment": "Move useClientAction into this package so every Vue host shares one write path, and file an action's returned record under the identity the server settled on rather than the id that was dispatched.",
+      "type": "minor",
+      "packageName": "@stonecrop/stonecrop"
+    }
+  ],
+  "packageName": "@stonecrop/stonecrop"
+}
