+++ docs/reference/desktop.md
@@ -9,6 +9,14 @@ description: Desktop navigation and command palette
 
 ## Vue Components
 
+### ActionEventPayload
+
+Vue component exported from @stonecrop/desktop.
+
+```typescript
+import { ActionEventPayload } from '@stonecrop/desktop'
+```
+
 ### ActionSet
 
 Vue component exported from @stonecrop/desktop.
@@ -59,21 +67,6 @@ Superset of all element types in the Action Set
 
 ```typescript
 export type ActionElements = ButtonElement | DropdownElement;
-```
-
-### ActionEventPayload
-
-Payload emitted with the 'action' event when the user triggers an FSM transition
-
-**Definition:**
-
-```typescript
-export type ActionEventPayload = {
-    name: string;
-    doctype: string;
-    recordId: string;
-    data: Record<string, any>;
-};
 ```
 
 ### BaseElement
