+++ nuxt/docs/decisions/0001-clienthandler-execution-host-delegation.md
@@ -50,6 +50,13 @@ Concretely: a pure `executeClientHandler(code, api)` lives in `@stonecrop/stonec
 * Bad, because it bakes the change into a `dist`-consumed package (rebuild, version bump, every consumer affected) and is hard to reverse.
 * Neutral, because Desktop deliberately routes through a `routeAdapter`, so it has no clean concrete `router` to inject anyway — the would-be locality advantage is smaller than it looks.
 
+## History
+
+* **2026-06-25** — decided (Option B).
+* **2026-08-11** — the composable **moved to `@stonecrop/stonecrop`**; `@stonecrop/nuxt` re-exports it so hosts and the auto-import are unchanged. The decision itself stands — Desktop is still a pure emitter, and the host still binds `@action` — but this ADR's stated reason for putting the composable in the *Nuxt* package does not: it said the four capabilities "cannot be assembled by any single package except at the host", and on measurement three come from Stonecrop while `vue-router` was already a `@stonecrop/stonecrop` peerDependency, imported in `registry.ts`. Nothing in it was Nuxt-specific. The cost of the mistake was a two-tier split: Nuxt hosts got the blessed write path in one line, every other Vue host re-derived ~60 lines, and **both that did — `examples/desktop` and the FAB consumer — wrote the result back under the id they dispatched**, which is wrong for exactly the created record. Three overrides (`buildArgs`, `followRecord`, `onError`) now cover what genuinely varies; identity resolution stays sealed.
+* **2026-08-11** — the store write moved one layer further down, into `Stonecrop.dispatchAction`, so a host that never adopts the composable still cannot file a record under the wrong key. The composable keeps only what needs the *dispatched* id — dropping the stale key and following the route — because that id lives inside the opaque `args` array, which the lower layer must not parse.
+* **2026-08-11** — the "module-provided default handler" deferred below is **no longer blocked**: it was waiting on "a host-provided fetch contract", and `DataClient` became exactly that. It stays deferred on its own merits — Desktop has no `vue-router` (it routes through `routeAdapter`, whose `NavigationTarget` cannot express `replace`), and all four hosts bind `@action` today, so a Desktop-side default would silently double-write until each migrated.
+
 ## More Information
 
 Executor: `stonecrop/src/client-handler.ts`. Composable: `nuxt/src/runtime/app/composables/useClientAction.ts`. Host wire: `nuxt/fullstack/app/pages/index.vue`. Scope is **transitions only**; the Command render path is deferred (Desktop's `actionElements` renders only FSM transitions today). A "module-provided default handler" that would remove even the one-line host binding is deferred: it requires a `<StonecropDesktop>` wrapper that also owns `@load-record`/`@load-records`, which are host-specific data fetches, so it needs a host-provided fetch contract. Capability details live in [0002](0002-clienthandler-capability-contract.md); the test approach in [0003](0003-clienthandler-execution-test-strategy.md).
