+++ stonecrop/src/types/index.ts
@@ -1,4 +1,5 @@
 // Re-export all type files
+export * from './client-action'
 export * from './composable'
 export * from './doctype'
 export * from './field-triggers'
